move test_vnet_decap to a common utils.
